Simpler and safe implementation with improved UX (#2)
* Simpler and safe implementation with improved UX

This is essentially two changes bundled into one (and I'll explain in
just a second why I bundled them)

The first change was to remove the dependence on `unsafeCoerce` by
changing the internal implementation to use an implicit parameter.
In fact, this is literally the exact same way `HasCallStack` is
implemented except without the GHC magic to automatically satisfy the
constraint if it is omitted.  In other words, all that the
`require-callstack` package does is implement `HasCallStack` without
the magic so that the constraint has to be explicitly propagated.

One of the benefits of this approach is that now users can also satisfy
the constraint by just defining the missing implicit parameter:

```haskell
let ?provideCallStack = ()
```

… but I kept the existing `provideCallStack` function both for
backwards compatibility and also for the benefit of users who don't
want to enable `ImplicitParams` (which is probably a lot of them, to be
honest).

The second change was to switch to using `TypeError` to improve the UX
instead of a cleverly named implicit parameter like
`?add_RequireCallStack_to_function_context_or_use_provideCallStack`.
The reason I bundled this change was because the choice of whether or
not to use `TypeError` affects what to name the implicit parameter.
I wanted to avoid having to rename the implicit parameter if we were
to switch to `TypeError` later since the implicit parameter's name
would technically become part of our public API if we were to switch to
this new approach (since users could begin defining the implicit
parameter instead of using `provideCallStack`).

So I did some research to figure out if it was feasible to switch to
`TypeError` right now and the answer is: yes!  Specifically, I tested
both `TypeError` and `Unsatisfiable` and learned the following things:

- `TypeError` actually works fine if you add type annotations to all
  intermediate `let` helper bindings that wrap `errorRequiredCallStack`

- GHC 9.8's support for `Unsatisfiable` doesn't actually change the
  design space at all

  Both `TypeError` and `Unsatisfiable` have the exact same problem
  where they trigger premature type errors if you define intermediate
  `let` helper bindings without a type annotation.

So since 9.8 isn't going to change the design space, I decided to see
if it was realistic to just ask people to annotate intermediate `let`
bindings that use `errorCallStack`.  Specifically, I tested this PR
against our internal codebase and found that it required minimal
changes to adopt this new `TypeError` approach.

This change only required adding 11 type annotations across 3 modules
(out of 6000+ modules), so I inferred from this that it wouldn't be
significantly disruptive to end users to switch to this new
`TypeError`-based UX.

Moreover, this change actually slightly improved GHC's constraint
resolution.  After making these two changes GHC now found two
redundant `HasCallStack` constraints within our codebase, whereas the
previous approach was somehow confounding GHC's ability to detect those
gratuitous constraints.

* Require users to use `provideCallStack`

… as suggested by @parsonsmatt

* Rename `CallStack` to `ProvideCallStack`

… based on a suggestion from @hw202207